Renamed "utils" to "utility" to align with other package names
