+++ stable-toolkit-back/src/main/java/it/demib/stabletoolkitback/service/IndexService.java
@@ -3,9 +3,9 @@ package it.demib.stabletoolkitback.service;
 import it.demib.stabletoolkitback.constant.ImageType;
 import it.demib.stabletoolkitback.model.entity.Folder;
 import it.demib.stabletoolkitback.model.entity.Image;
-import it.demib.stabletoolkitback.utils.GenerationInformationTextParser;
-import it.demib.stabletoolkitback.utils.PNGTextExtractor;
-import it.demib.stabletoolkitback.utils.TagParser;
+import it.demib.stabletoolkitback.utility.GenerationInformationTextParser;
+import it.demib.stabletoolkitback.utility.PNGTextExtractor;
+import it.demib.stabletoolkitback.utility.TagParser;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.FileInputStream;
